fix(shelter): shelter api shelter app내부로 이동
