+++ packages/shared/apis/volunteers/Volunteer.ts
@@ -1,17 +1,5 @@
 import axiosInstance from '../axiosInstance';
 
-type SignUpParams = {
-  email: string;
-  password: string;
-  name: string;
-  birthDate: string;
-  phoneNumber: string;
-  gender: 'FEMALE' | 'MALE';
-};
-
-export const signUpVolunteer = (signUpParams: SignUpParams) =>
-  axiosInstance.post<unknown, SignUpParams>('/volunteers', signUpParams);
-
 type MyInfoResponse = {
   volunteerId: string;
   volunteerEmail: string;
@@ -70,12 +58,3 @@ type ApplicantsResponse = {
 //봉사자가 신청한 봉사 리스트 조회
 export const getApplicants = () =>
   axiosInstance.get<ApplicantsResponse>('/volunteers/applicants');
-
-export const checkDuplicatedVolunteerEmail = (email: string) => {
-  return axiosInstance.post<{ isDuplicated: boolean }, { email: string }>(
-    '/volunteers/email',
-    {
-      email,
-    },
-  );
-};
